Allow JSON params to be supplied via stdin
